Added functions that loop over MODEL_res.tbl files and sort the results by protein.
This is currently done in process_models.py and will need to be moved
into an existing or new class (might be better as this can be done
separately).

I will need to rework the data_[0-9].pkl files to be dictionaries of
genomes with lists of proteins as this will be the desired layout for
scoring genomes later (will also help with the small genomes check
step).